exercicio

+++ src/Trabalho_Etapa2_POO_ClaudimarCruz/Model/FuncionarioModel.java
@@ -3,24 +3,24 @@ package Trabalho_Etapa2_POO_ClaudimarCruz.Model;
 import java.util.List;
 
 public class FuncionarioModel extends PessoaModel {
-	private Double salario;
+	private Float salario;
 	private String cargo;
 	
 	public FuncionarioModel() {
 		super();
 	}
 
-	public FuncionarioModel(Integer codigo, String nome, String cpf, Double salario, String cargo, String endereco, List<String> telefones) {
+	public FuncionarioModel(Integer codigo, String nome, String cpf, Float salario, String cargo, String endereco, List<String> telefones) {
 		super(codigo, nome, cpf, endereco, telefones);
 		this.salario = salario;
 		this.cargo = cargo;
 	}
 
-	public Double getSalario() {
+	public Float getSalario() {
 		return salario;
 	}
 
-	public void setSalario(Double salario) {
+	public void setSalario(Float salario) {
 		this.salario = salario;
 	}
